requirements.txt
django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python_dotenv==1.0.1
